add cell center

+++ graphics.py
@@ -2,7 +2,7 @@ from tkinter import Tk, BOTH, Canvas
 
 
 class Point:
-    def __init__(self, x: int, y: int) -> None:
+    def __init__(self, x: float, y: float) -> None:
         self.x = x
         self.y = y
